requirements.txt
llama-index-core>=0.10.0
llama-index-llms-llama-cpp>=0.1.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-vector-stores-chroma>=0.1.0
chromadb>=0.4.13
pypdf>=3.1.0
chardet>=5.0.0
transformers>=4.36.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
requests>=2.28.0

# Use CPU-only torch
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
torchvision>=0.15.0
